requirements.txt
tensorflow==1.7
scipy
scikit-learn
scikit-image
opencv-python
h5py
matplotlib
Pillow
requests
psutil
torch
torchvision
